Decouple TaskList from Storage class
